// server/src/data/sevenSignsDungeonMobSpawns.generated.ts
export interface SevenSignsDungeonMobSpawn {
  id: string;
  dungeonId: string;
  mapX: number;
  mapY: number;
  npcId: number;
  name: string;
  level: number;
  aggressive: boolean;
}